+++ sandbox/src/oanda/tao_bot/set_status.cpp
@@ -18,7 +18,7 @@
 #include <iostream>                // std::cout, std::endl
 #include <math.h>                  // abs
 #include <stdio.h>                 // printf
-#include <string>                  // std::string, std::stoi
+#include <string>                  // std::string, std::stol
 
 void Oanda::TaoBot::set_status(order_t *order,
                                order_t *linked_order = nullptr) {
@@ -79,7 +79,7 @@ void Oanda::TaoBot::set_status(order_t *order,
       order->status == order_status_t::ORDER_FILLED) {
     const char *order_action = Oanda::constants::ORDER_ACTIONS[order->action];
     const char *log_icon = this->ICONS[order_action];
-    const int trade_quantity = abs(std::stoi(trade_quantity_string));
+    const long int trade_quantity = abs(std::stol(trade_quantity_string));
 
     std::cout << fmt.bold << fmt.green << std::endl;
     printf("%s Executed %s order.\n", log_icon, order_action);
@@ -87,8 +87,8 @@ void Oanda::TaoBot::set_status(order_t *order,
 
     if (trade_quantity != order->quantity) {
       std::cout << fmt.bold << fmt.red << std::endl;
-      printf("🚨 Executed trade quantity (%i) doesn't match desired quantity "
-             "(%i). Adjusting.\n",
+      printf("🚨 Executed trade quantity (%li) doesn't match desired quantity "
+             "(%li). Adjusting.\n",
              trade_quantity, order->quantity);
       std::cout << fmt.reset;
